# habitly-export/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Habitly API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid data
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /coach/messages:
    get:
      operationId: getCoachMessages
      tags: [coach]
      summary: Get coach conversation
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachMessage"
    post:
      operationId: sendCoachMessage
      tags: [coach]
      summary: Send a message to the AI coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Coach reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachMessage"
...
  /dashboard/adaptations/{habitId}/apply:
    post:
      operationId: applyAdaptation
      tags: [dashboard]
      summary: Apply a suggested adaptation to a habit
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyAdaptationBody"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        tier:
          type: string
          enum: [free, premium]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - tier
        - createdAt
    SignupBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
      required:
        - email
        - name
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        tier:
          type: string
          enum: [free, premium]
...
    Habit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        purpose:
          type: string
        frequency:
          type: string
          enum: [daily, weekdays, weekends, custom]
        durationMinutes:
          type: integer
        bestTimeOfDay:
          type: string
          enum: [morning, afternoon, evening, anytime]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        fallbackMicroHabit:
          type: string
        isActive:
          type: boolean
        currentStreak:
          type: integer
        completionRate:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - purpose
        - frequency
        - durationMinutes
        - bestTimeOfDay
        - difficulty
        - fallbackMicroHabit
        - isActive
        - currentStreak
        - createdAt
        - updatedAt
...
    CoachMessage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        mood:
          type: ["integer", "null"]
        blockerNote:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - role
        - content
        - createdAt
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        mood:
          type: ["integer", "null"]
        blockerNote:
          type: ["string", "null"]
      required:
        - content
...
    ApplyAdaptationBody:
      type: object
      properties:
        type:
          type: string
        changes:
          type: object
      required:
        - type
        - changes